requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
cryptography==39.0.0
Deprecated==1.2.13
Django==4.1.2
greenlet==1.1.3.post0
humanize==4.4.0
logging-tree==1.9
loguru==0.6.0
logx==0.1.1579232358
mariadb==1.1.5.post3
packaging==21.3
psycopg2==2.9.4
pika==1.3.1
psutil==5.9.4
pyparsing==3.0.9
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
redis==4.3.4
ruamel.yaml==0.17.21
six==1.16.0
SQLAlchemy==1.4.42
sqlparse==0.4.3
tzdata==2022.5
tzlocal==4.2
wrapt==1.14.1
dj-database-url==1.0.0
pyodbc==4.0.35